requirements.txt
fastapi
uvicorn
pymongo
motor
pytest
langchain
openai
faiss-cpu
python-dotenv